# swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Aircraft Training ETR API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Approvals": {
      "get": {
        "tags": [
          "Approvals"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Approvals"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApprovalRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApprovalRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApprovalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Approvals/{id}/process": {
      "post": {
        "tags": [
          "Approvals"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "action",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "comment",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Enrollments": {
      "get": {
        "tags": [
          "Enrollments"
        ],
        "summary": "[Module/Flow]: ETR Processing\r\n[Core Responsibility]: Retrieves all course enrollments.\r\n[Target Audience]: Admin",
        "responses": {
          "200": {
            "description": "Returns the list of enrollments.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnrollmentResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnrollmentResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnrollmentResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "If the user is not authenticated."
          },
          "403": {
            "description": "If the user is not an Admin."
          }
        }
      },
      "post": {
        "tags": [
          "Enrollments"
        ],
        "summary": "Creates a new course enrollment for a learner.",
        "requestBody": {
          "description": "The enrollment details.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEnrollmentRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEnrollmentRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEnrollmentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns the newly created enrollment."
          },
          "401": {
            "description": "If the user is not authenticated."
          },
          "403": {
            "description": "If the user is not an Admin."
          }
        }
      }
    },
    "/api/Enrollments/{id}": {
      "get": {
        "tags": [
          "Enrollments"
        ],
        "summary": "[Module/Flow]: ETR Processing\r\n[Core Responsibility]: Retrieves a specific course enrollment by ID.\r\n[Target Audience]: Admin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The Enrollment ID.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the enrollment details.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentResponse"
                }
              }
            }
          },
          "401": {
            "description": "If the user is not authenticated."
          },
          "403": {
            "description": "If the user is not an Admin."
          },
          "404": {
            "description": "If the enrollment is not found."
          }
        }
      }
    },
    "/api/Enrollments/student/{studentId}": {
      "get": {
        "tags": [
          "Enrollments"
        ],
        "summary": "Retrieves enrollments for a specific student.",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnrollmentResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnrollmentResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnrollmentResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Evidences": {
      "get": {
        "tags": [
          "Evidences"
        ],
        "summary": "Retrieves all uploaded evidence files.",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Evidences"
        ],
        "summary": "Uploads a new evidence file record.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEvidenceRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEvidenceRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEvidenceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Evidences/{id}": {
      "get": {
        "tags": [
          "Evidences"
        ],
        "summary": "Retrieves a specific evidence file by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Evidences"
        ],
        "summary": "Soft deletes an evidence file.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateApprovalRequest": {
        "type": "object",
        "properties": {
          "etrCourseRecordId": {
            "type": "integer",
            "format": "int32"
          },
          "submittedBy": {
            "type": "integer",
            "format": "int32"
          },
          "currentApproverId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateEnrollmentRequest": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "integer",
            "format": "int32"
          },
          "classId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CreateEvidenceRequest": {
        "required": [
          "accountId",
          "evidenceTypeId",
          "fileName",
          "filePath",
          "fileSize",
          "subjectResultId"
        ],
        "type": "object",
        "properties": {
          "evidenceTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "accountId": {
            "type": "integer",
            "format": "int32"
          },
          "subjectResultId": {
            "type": "integer",
            "format": "int32"
          },
          "attendanceRecordId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "assessmentResultId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "fileName": {
            "minLength": 1,
            "type": "string"
          },
          "filePath": {
            "minLength": 1,
            "type": "string"
          },
          "fileExtension": {
            "type": "string",
            "nullable": true
          },
          "mimeType": {
            "type": "string",
            "nullable": true
          },
          "fileSize": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
...
      "EnrollmentResponse": {
        "type": "object",
        "properties": {
          "enrollmentId": {
            "type": "integer",
            "format": "int32"
          },
          "accountId": {
            "type": "integer",
            "format": "int32"
          },
          "classId": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "enrolledAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Nhập token của bạn vào đây (KHÔNG CẦN gõ chữ 'Bearer ' ở trước).",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}